+++ bundles/org.openhab.binding.groupepsa/src/main/java/org/openhab/binding/groupepsa/internal/rest/api/dto/Odometer.java
@@ -12,6 +12,7 @@
  */
 package org.openhab.binding.groupepsa.internal.rest.api.dto;
 
+import java.math.BigDecimal;
 import java.time.ZonedDateTime;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
@@ -24,13 +25,13 @@ import org.eclipse.jdt.annotation.Nullable;
 public class Odometer {
 
     private @Nullable ZonedDateTime createdAt;
-    private @Nullable Double mileage;
+    private @Nullable BigDecimal mileage;
 
     public @Nullable ZonedDateTime getCreatedAt() {
         return createdAt;
     }
 
-    public @Nullable Double getMileage() {
+    public @Nullable BigDecimal getMileage() {
         return mileage;
     }
