+++ bundles/org.openhab.binding.groupepsa/src/main/java/org/openhab/binding/groupepsa/internal/rest/api/dto/Properties.java
@@ -12,6 +12,8 @@
  */
 package org.openhab.binding.groupepsa.internal.rest.api.dto;
 
+import java.math.BigDecimal;
+
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
 
@@ -21,15 +23,15 @@ import org.eclipse.jdt.annotation.Nullable;
 @NonNullByDefault
 public class Properties {
 
-    private @Nullable Double heading;
-    private @Nullable Double signalQuality;
+    private @Nullable BigDecimal heading;
+    private @Nullable BigDecimal signalQuality;
     private @Nullable String type;
 
-    public @Nullable Double getHeading() {
+    public @Nullable BigDecimal getHeading() {
         return heading;
     }
 
-    public @Nullable Double getSignalQuality() {
+    public @Nullable BigDecimal getSignalQuality() {
         return signalQuality;
     }
